最新のAPIドキュメントをもとにモデルを更新
