+++ vpsc/models/generated.py
@@ -1,0 +1,490 @@
+# generated by datamodel-codegen:
+#   filename:  api-json.json
+#   timestamp: 2024-07-06T05:16:28+00:00
+# datamodel-codegen --input ./api-json.json --input-file-type openapi --output-model-type pydantic_v2.BaseModel --enum-field-as-literal all --reuse-model --strict-nullable --target-python-version 3.8 | sed -E "s/description='([^']+)'/description=\"\"\"\1\"\"\"/; s/\\\n/\n/g" > generated.py
+from __future__ import annotations
+
+from typing import List, Literal, Optional, Union
+
+from pydantic import AnyUrl, AwareDatetime, BaseModel, Field, RootModel, conint, constr
+
+
+class Pagination(BaseModel):
+    count: int = Field(..., description="""データ総数""", examples=[100])
+    next: Optional[AnyUrl] = Field(
+        ..., description="""次のページへのURL""", examples=["https://api.example.com/?page=3&per_page=10"]
+    )
+    previous: Optional[AnyUrl] = Field(
+        ..., description="""前のページへのURL""", examples=["https://api.example.com/?per_page=10"]
+    )
+
+
+class InvalidParameterDetailItem(BaseModel):
+    code: Optional[str] = Field(None, description="""エラー内容を示す簡潔な識別子""", examples=["required"])
+    message: Optional[str] = Field(None, examples=["この項目は必須です"])
+
+
+class InvalidParameterDetail(RootModel[List[InvalidParameterDetailItem]]):
+    root: List[InvalidParameterDetailItem]
+
+
+class Errors(BaseModel):
+    non_field_errors: Optional[InvalidParameterDetail] = None
+
+
+class ProblemDetails400(BaseModel):
+    code: Optional[Literal["invalid", "parse_error", "bad_request"]] = Field(
+        None,
+        description="""エラー内容を示す簡潔な識別子
+* `invalid` - 不正なリクエスト値,リクエスト値が妥当でない
+* `parse_error` - 不正な形式,リクエスト値を読み取ることができない
+* `bad_request` - リクエストの内容に何らかの問題がある""",
+    )
+    message: Optional[str] = Field(None, description="""エラーの内容""", examples=["Invalid input."])
+    errors: Optional[Errors] = Field(
+        None,
+        description="""入力値に対するエラーを構造化した情報
+(code`invalid`の場合のみ)
+* `non_field_errors` - リクエスト全体に起因した(単一項目でない)エラー内容
+* `*` - 対応した入力項目ごとのエラー内容""",
+        examples=[{"foo": [{"code": "required", "message": "この項目は必須です"}]}],
+    )
+
+
+class ProblemDetails404(BaseModel):
+    code: Optional[Literal["not_found"]] = None
+    message: Optional[str] = Field(None, description="""エラーの内容""", examples=["見つかりませんでした。"])
+
+
+class ProblemDetails409(BaseModel):
+    code: Optional[Literal["conflict"]] = None
+    message: Optional[str] = Field(None, description="""エラーの内容""", examples=["状態の競合によりリクエストを処理できません。"])
+
+
+class ProblemDetails429(BaseModel):
+    code: Optional[Literal["throttled"]] = None
+    message: Optional[str] = Field(
+        None, description="""エラーの内容""", examples=["リクエストの処理は絞られました。 Expected available in xxxxxxx."]
+    )
+
+
+class ProblemDetails503(BaseModel):
+    code: Optional[Literal["temporary_unavailable"]] = None
+    message: Optional[str] = Field(None, description="""エラーの内容""", examples=["一時的にご利用になれません。"])
+
+
+class Limitation(BaseModel):
+    cpu_performance_limit: Optional[Literal["enabled", "disabled"]] = Field(None, description="""CPUリソースの制限可否""")
+    network_bandwidth_limit: Optional[Literal["enabled", "disabled"]] = Field(None, description="""ネットワーク帯域の制限可否""")
+    outbound_port_25_blocking: Optional[Literal["enabled", "disabled"]] = Field(None, description="""OP25Bの可否""")
+    storage_iops_limit: Optional[Literal["enabled", "disabled"]] = Field(None, description="""ストレージのIOPSの制限可否""")
+
+
+class StorageItem(BaseModel):
+    port: int = Field(..., description="""ポート番号""", examples=[0])
+    type: Literal["ssd", "hdd"] = Field(..., description="""種別""")
+    size_gibibytes: int = Field(..., description="""ストレージ容量(GiB)""", examples=[100])
+
+
+class Zone(BaseModel):
+    code: Literal["tk1", "tk2", "tk3", "os1", "os2", "os3", "is1"] = Field(
+        ...,
+        description="""ゾーンコード
+* tk1 東京第1
+* tk2 東京第2
+* tk3 東京第3
+* os1 大阪第1
+* os2 大阪第2
+* os3 大阪第3
+* is1 石狩第1""",
+    )
+    name: str = Field(..., description="""ゾーン名称""", examples=["石狩第1"])
+
+
+class Ipv4(BaseModel):
+    address: str = Field(..., description="""アドレス""", examples=["198.51.100.2"])
+    netmask: str = Field(..., description="""サブネットマスク""", examples=["255.255.254.0"])
+    gateway: str = Field(..., description="""ゲートウェイのアドレス""", examples=["198.51.100.1"])
+    nameservers: List[str] = Field(..., description="""ネームサーバーのアドレスリスト""")
+    hostname: str = Field(..., description="""標準ホスト名""", examples=["example.jp"])
+    ptr: str = Field(..., description="""逆引きホスト名""", examples=["example.jp"])
+
+
+class Ipv6(BaseModel):
+    address: Optional[str] = Field(..., description="""アドレス""", examples=["2001:e42:102:1501:153:121:89:107"])
+    prefixlen: Optional[int] = Field(..., description="""プレフィックス長""", examples=[64])
+    gateway: Optional[str] = Field(..., description="""ゲートウェイのアドレス""", examples=["fe80::1"])
+    nameservers: List[str] = Field(..., description="""ネームサーバーのアドレスリスト""")
+    hostname: Optional[str] = Field(..., description="""標準ホスト名""", examples=["example.jp"])
+    ptr: Optional[str] = Field(..., description="""逆引きホスト名""", examples=["example.jp"])
+
+
+class Contract(BaseModel):
+    plan_code: int = Field(..., description="""プランコード""", examples=[3439])
+    plan_name: str = Field(..., description="""プラン名""", examples=["さくらのVPS(v5)  1G IK01"])
+    service_code: str = Field(..., description="""サービスコード""", examples=["100000000000"])
+
+
+class Server(BaseModel):
+    id: int = Field(..., description="""id""")
+    name: str = Field(..., description="""名前""")
+    description: str = Field(..., description="""説明""")
+    service_type: Literal["linux", "windows"] = Field(..., description="""サービスタイプ""")
+    service_status: Literal["on_trial", "link_down_on_trial", "in_use", "link_down"] = Field(
+        ...,
+        description="""サービス状況
+* on_trial お試し中
+* link_down_on_trial お試し中（一時停止）
+* in_use 利用中
+* link_down 一時停止中""",
+    )
+    cpu_cores: int = Field(..., description="""CPUコア数""", examples=[2])
+    memory_mebibytes: int = Field(..., description="""メモリ容量(MiB)""", examples=[1024])
+    storage: List[StorageItem] = Field(..., description="""ストレージ情報""")
+    zone: Zone = Field(..., description="""ゾーン情報""")
+    options: List[str] = Field(..., description="""オプション（追加ソフトウェア）""")
+    version: str = Field(..., description="""プランの世代""", examples=["v5"])
+    ipv4: Ipv4
+    ipv6: Ipv6
+    contract: Contract = Field(..., description="""契約情報""")
+    power_status: Literal[
+        "power_on", "in_shutdown", "power_off", "installing", "in_scaleup", "migration", "unknown"
+    ] = Field(
+        ...,
+        description="""電源ステータス
+* power_on 電源ON
+* in_shutdown シャットダウン中
+* power_off 電源OFF
+* installing OSインストール中
+* in_scaleup スケールアップ中
+* migration サーバー移行作業中
+* unknown 不明（電源状態を取得できない）
+このエンドポイントが返す電源ステータスはキャッシュされた情報のため、最新の正確な電源ステータスが必要な場合は
+**サーバーの電源状態を取得する**`/servers/{server_id}/power-status`をご利用ください""",
+    )
+
+
+class ServerPowerStatus(BaseModel):
+    status: Literal[
+        "power_on", "in_shutdown", "power_off", "installing", "in_scaleup", "migration", "unknown"
+    ] = Field(
+        ...,
+        description="""電源ステータス
+* power_on 電源ON
+* in_shutdown シャットダウン中
+* power_off 電源OFF
+* installing OSインストール中
+* in_scaleup スケールアップ中
+* migration サーバー移行作業中
+* unknown 不明（電源状態を取得できない）""",
+    )
+
+
+class ServerInterface(BaseModel):
+    id: int = Field(..., description="""id""")
+    device: str = Field(..., description="""NIC名称""", examples=["eth0"])
+    connectable_to_global_network: bool = Field(..., description="""グローバルネットワークと接続可能か""", examples=[True])
+    connect_to: Optional[Literal["switch", "global"]] = Field(..., description="""インターフェースの接続先""", examples=["global"])
+    mac: str = Field(..., description="""MACアドレス""", examples=["9C:A3:BA:00:00:00"])
+    switch_id: Optional[int] = Field(..., description="""スイッチID""", examples=[1])
+
+
+class StorageItem1(BaseModel):
+    type: Literal["ssd", "hdd"] = Field(..., description="""種別""")
+    size_gibibytes: int = Field(..., description="""ストレージ容量(GiB)""", examples=[100])
+
+
+class Ipv41(BaseModel):
+    address: str = Field(..., description="""アドレス""", examples=["198.51.100.2"])
+    netmask: str = Field(..., description="""サブネットマスク""", examples=["255.255.254.0"])
+
+
+class Contract1(BaseModel):
+    plan_code: int = Field(..., description="""プランコード""", examples=[3439])
+    plan_name: str = Field(..., description="""プラン名""", examples=["さくらのVPS(v5)  NFS 200GB OS03"])
+    service_code: str = Field(..., description="""サービスコード""", examples=["100000000000"])
+
+
+class NfsServer(BaseModel):
+    id: int = Field(..., description="""id""")
+    name: str = Field(..., description="""名前""")
+    description: str = Field(..., description="""説明""")
+    service_status: Literal["in_preparation", "on_trial", "link_down_on_trial", "in_use", "link_down"] = Field(
+        ...,
+        description="""サービス状況
+* in_preparation 準備中
+* on_trial お試し中
+* link_down_on_trial お試し中（一時停止）
+* in_use 利用中
+* link_down 一時停止中""",
+    )
+    setting_status: Literal["done", "in_update", "failed"] = Field(
+        ...,
+        description="""設定状況
+* done 設定完了
+* in_update 設定更新中
+* failed 設定更新失敗""",
+    )
+    storage: List[StorageItem1] = Field(..., description="""ストレージ情報""")
+    zone: Zone = Field(..., description="""ゾーン情報""")
+    ipv4: Ipv41
+    contract: Contract1 = Field(..., description="""契約情報""")
+    power_status: Literal["power_on", "in_shutdown", "power_off", "unknown"] = Field(
+        ...,
+        description="""電源ステータス
+* power_on 電源ON
+* in_shutdown シャットダウン中
+* power_off 電源OFF
+* unknown 不明（電源状態を取得できない）
+このエンドポイントが返す電源ステータスはキャッシュされた情報のため、最新の正確な電源ステータスではない場合があります""",
+    )
+
+
+class NfsServerPowerStatus(BaseModel):
+    status: Literal["power_on", "in_shutdown", "power_off", "unknown"] = Field(
+        ...,
+        description="""電源ステータス
+* power_on 電源ON
+* in_shutdown シャットダウン中
+* power_off 電源OFF
+* unknown 不明（電源状態を取得できない）""",
+    )
+
+
+class NfsServerInterface(BaseModel):
+    id: int = Field(..., description="""id""")
+    device: str = Field(..., description="""NIC名称""", examples=["eth0"])
+    connect_to: Optional[Literal["switch"]] = Field(..., description="""インターフェースの接続先""", examples=["switch"])
+    mac: str = Field(..., description="""MACアドレス""", examples=["9C:A3:BA:00:00:00"])
+    switch_id: Optional[int] = Field(..., description="""スイッチID""", examples=[1])
+
+
+class Service(BaseModel):
+    service_category: str = Field(..., description="""サービスカテゴリー""", examples=["cloud"])
+    service_name: str = Field(..., description="""サービス名""", examples=["クラウド東京第1ゾーン"])
+    switch_code: str = Field(..., description="""スイッチコード""", examples=["111111111111"])
+
+
+class ExternalConnection(BaseModel):
+    service_code: str = Field(..., description="""サービスコード""", examples=["100000000000"])
+    type: Literal["cloud", "sales", "localrouter", "awsdxcon"] = Field(..., description="""外部接続方式""")
+    services: List[Service]
+
+
+class Switch(BaseModel):
+    id: int = Field(..., description="""id""")
+    name: str = Field(..., description="""名前""")
+    description: str = Field(..., description="""説明""")
+    switch_code: str = Field(..., description="""スイッチコード""")
+    zone: Zone = Field(..., description="""ゾーン情報""")
+    server_interfaces: List[int] = Field(..., description="""接続されているサーバーのインターフェースid""")
+    nfs_server_interfaces: List[int] = Field(..., description="""接続されている追加ストレージ（NFS）のインターフェースid""")
+    external_connection: Optional[ExternalConnection] = Field(..., description="""接続されている外部接続の情報""")
+
+
+class Disc(BaseModel):
+    id: int = Field(..., description="""id""")
+    name: str = Field(..., description="""名前""")
+    description: str = Field(..., description="""説明""")
+    license_required: bool = Field(..., description="""ライセンスが必要かどうか""")
+
+
+class Email(BaseModel):
+    enabled: bool = Field(
+        ...,
+        description="""通知のON/OFF
+* true 通知ON
+* false 通知OFF""",
+    )
+
+
+class IncomingWebhook(BaseModel):
+    enabled: bool = Field(
+        ...,
+        description="""通知のON/OFF
+* true 通知ON
+* false 通知OFF""",
+    )
+    webhooks_url: Optional[AnyUrl] = Field(
+        ...,
+        description="""通知先のWebhooksURL。
+Slack、Discord、Microsoft TeamsのIncoming WebHooksにのみ対応しています。
+指定できるURLは各サービスのWebhook URL(https://hooks.slack.com/services/* など)の形式に制限されています。
+Discordの場合は[Slack互換のWebhook URL](https://discord.com/developers/docs/resources/webhook#execute-slackcompatible-webhook)を指定してください。""",
+    )
+    slack_team_name: str = Field(..., description="""slackのteam name。VPSコントロールパネルの「Slackと自動で連携をする」を利用した場合に設定されます。""")
+    slack_channel_name: str = Field(
+        ..., description="""slackのchannel name。VPSコントロールパネルの「Slackと自動で連携をする」を利用した場合に設定されます。"""
+    )
+
+
+class Notification(BaseModel):
+    email: Email = Field(..., description="""emailでの通知の設定。会員情報に登録されているメールアドレス宛に送信されます。""")
+    incoming_webhook: IncomingWebhook = Field(..., description="""incoming webhookでの通知の設定""")
+    interval_hours: conint(ge=1, le=72) = Field(..., description="""再通知間隔(時間)""")
+
+
+class HealthCheckPing(BaseModel):
+    protocol: Literal["ping"] = Field(..., description="""監視方法""")
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckTcp(BaseModel):
+    protocol: Literal["tcp"] = Field(..., description="""監視方法""")
+    port: conint(ge=1, le=65535) = Field(..., description="""ポート番号""")
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckSsh(BaseModel):
+    protocol: Literal["ssh"] = Field(..., description="""監視方法""")
+    port: conint(ge=1, le=65535) = Field(..., description="""ポート番号""", examples=[22])
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckSmtp(BaseModel):
+    protocol: Literal["smtp"] = Field(..., description="""監視方法""")
+    port: conint(ge=1, le=65535) = Field(..., description="""ポート番号""", examples=[25])
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckPop3(BaseModel):
+    protocol: Literal["pop3"] = Field(..., description="""監視方法""")
+    port: conint(ge=1, le=65535) = Field(..., description="""ポート番号""", examples=[110])
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckHttpBase(BaseModel):
+    port: conint(ge=1, le=65535) = Field(..., description="""ポート番号""", examples=[80])
+    host: Optional[constr(min_length=1, max_length=63)] = Field(
+        ...,
+        description="""監視用HTTPリクエストのHostヘッダ  
+RFCの定義に基づいて下記の制限をかけています
+* ラベルは2つ以上必要
+* 各ラベルについて
+  * 利用できる文字
+    * 半角数字 0～9
+    * 半角英小文字 a～z
+    * 半角記号 -
+  * 両端が-でないこと
+  * -が2つ以上続かないこと。ただしpunycodeの接頭辞`xn--`を除く
+  * 下記パターン(RFCなどで予約または禁止されているパターン)と一致しないこと
+    * isatap
+    * wpad
+    * example
+    * example0～example9
+* 最後のラベルについて
+  * 利用できる文字
+    * 半角英小文字 a～z
+  * 下記のパターン(RFC予約済みのDNS名)と一致しないこと
+    * test
+    * localhost""",
+    )
+    path: constr(min_length=1, max_length=1000) = Field(
+        ...,
+        description="""監視対象のパス
+* 利用できる文字 
+  * 半角数字 0～9
+  * 半角英字 A～Z、a～z
+  * 半角記号 _./~%?=-&""",
+        examples=["/"],
+    )
+    basic_auth_username: Optional[constr(min_length=1, max_length=20)] = Field(
+        ...,
+        description="""ベーシック認証のユーザー名
+* 利用できる文字 
+  * 半角数字 0～9
+  * 半角英字 A～Z、a～z
+  * 半角記号 _.-+!@""",
+    )
+    basic_auth_password: Optional[constr(min_length=1, max_length=32)] = Field(
+        ...,
+        description="""ベーシック認証のパスワード
+* 利用できる文字 
+  * 半角数字 0～9
+  * 半角英字 A～Z、a～z
+  * 半角記号 !#$%&()*+,-./:<=>?@[]^_`{|}~""",
+    )
+    status: conint(ge=100, le=999) = Field(..., description="""正常と見なすHTTPステータスコード""")
+    interval_minutes: conint(ge=1, le=60) = Field(..., description="""チェック間隔(分)""")
+
+
+class HealthCheckHttp(HealthCheckHttpBase):
+    protocol: Literal["http"] = Field(..., description="""監視方法""")
+
+
+class HealthCheckHttps(HealthCheckHttpBase):
+    protocol: Literal["https"] = Field(..., description="""監視方法""")
+    sni: bool = Field(
+        ...,
+        description="""SNIを設定しているWebサーバか
+* true SNI設定あり
+* false SNI設定なし""",
+    )
+
+
+class ServerMonitoringHealth(BaseModel):
+    last_checked_at: Optional[AwareDatetime] = Field(
+        ..., description="""最終監視日時""", examples=["2024-04-15T23:59:59+09:00"]
+    )
+    status: Optional[Literal["healthy", "unhealthy"]] = Field(
+        ...,
+        description="""ステータス
+  * healthy 正常
+  * unhealthy 異常""",
+    )
+    last_status_changed_at: Optional[AwareDatetime] = Field(
+        ..., description="""最終ステータス変更日時""", examples=["2024-04-01T23:59:59+09:00"]
+    )
+
+
+class Zone3(BaseModel):
+    id: int = Field(..., description="""id""")
+    code: Literal["tk1", "tk2", "tk3", "os1", "os2", "os3", "is1"] = Field(
+        ...,
+        description="""ゾーンコード
+* tk1 東京第1
+* tk2 東京第2
+* tk3 東京第3
+* os1 大阪第1
+* os2 大阪第2
+* os3 大阪第3
+* is1 石狩第1""",
+    )
+    name: str = Field(..., description="""ゾーン名称""")
+    can_use_local: bool = Field(..., description="""ローカルネットワーク接続が可能かどうか""")
+    can_use_hybrid: bool = Field(..., description="""ハイブリッド接続が利用可能かどうか""")
+
+
+class Settings(BaseModel):
+    enabled: bool = Field(
+        ...,
+        description="""監視のON/OFF
+* true 監視ON
+* false 監視OFF""",
+    )
+    health_check: Union[
+        HealthCheckPing,
+        HealthCheckTcp,
+        HealthCheckSsh,
+        HealthCheckSmtp,
+        HealthCheckPop3,
+        HealthCheckHttp,
+        HealthCheckHttps,
+    ]
+    notification: Notification
+
+
+class ServerMonitoring(BaseModel):
+    id: int = Field(..., description="""id""")
+    name: constr(max_length=255) = Field(..., description="""名前""")
+    description: constr(max_length=10000) = Field(..., description="""説明""")
+    monitoring_resource_id: str = Field(..., description="""監視リソースID""")
+    update_status: Literal["waiting", "updating", "completed", "error"] = Field(
+        ...,
+        description="""更新ステータス
+* waiting 更新待ち
+* updating 更新中
+* completed 更新完了
+* error 更新エラー""",
+    )
+    settings: Settings
